fix(struct): change gross amount data type to float64

+++ snap.go
@@ -24,7 +24,7 @@ func (gateway *SnapGateway) Call(method, path string, body io.Reader, v interfac
 }
 
 // GetTokenQuick : Quickly get token without constructing the body manually
-func (gateway *SnapGateway) GetTokenQuick(orderID string, grossAmount int64) (SnapResponse, error) {
+func (gateway *SnapGateway) GetTokenQuick(orderID string, grossAmount float64) (SnapResponse, error) {
 	return gateway.GetToken(&SnapReq{
 		TransactionDetails: TransactionDetails{
 			OrderID:  orderID,
@@ -52,7 +52,7 @@ func (gateway *SnapGateway) GetToken(r *SnapReq) (SnapResponse, error) {
 }
 
 // GetTokenQuickWithMap : Quickly get token without constructing the body manually
-func (gateway *SnapGateway) GetTokenQuickWithMap(orderID string, grossAmount int64) (ResponseWithMap, error) {
+func (gateway *SnapGateway) GetTokenQuickWithMap(orderID string, grossAmount float64) (ResponseWithMap, error) {
 	return gateway.GetTokenWithMap(&SnapReqWithMap{
 		"transaction_details": TransactionDetails{
 			OrderID:  orderID,
